+++ bzt/__init__.py
@@ -17,7 +17,7 @@ import os
 import sys
 from abc import abstractmethod
 
-VERSION = "1.6.7"
+VERSION = "1.6.8"
 
 
 class RCProvider(object):