+++ java/JUnit_plugin/src/main/java/taurusjunit/CustomListener.java
@@ -36,9 +36,8 @@ public class CustomListener extends RunListener {
         started = System.currentTimeMillis();
         pendingSample = new Sample();
         pendingSample.setLabel(description.getMethodName());
+        pendingSample.setSuite(description.getClassName());
         pendingSample.setFullName(description.getClassName() + "." + description.getMethodName());
-        pendingSample.setFile("");
-        pendingSample.setDescription("");
         testCount += 1;
     }