+++ site/dat/docs/Changelog.md
@@ -3,6 +3,7 @@
 ## 1.6.8
  - fix hamcrest installation for Java-based Selenium tests
  - fix Jmeter fail with unusual jmx property type
+ - fix results time and add widget for ApacheBenchmark
 
 ## 1.6.7  <sup>23 aug 2016</sup>
  - add `cloud` and `local` aliases