+++ site/dat/docs/Changelog.md
@@ -5,6 +5,7 @@
  - fix Jmeter fail with unusual jmx property type
  - add variable sending example into Gatling documentation
  - fix results time and add widget for ApacheBenchmark
+ - add `script: <jar-file>` capability to Gatling executor
 
 ## 1.6.7  <sup>23 aug 2016</sup>
  - add `cloud` and `local` aliases