+++ site/dat/docs/Changelog.md
@@ -1,13 +1,14 @@
 # Changelog
 
-## 1.6.8
+## 1.6.8 <sup>1 sep 2016</sup>
  - fix hamcrest installation for Java-based Selenium tests
  - fix Jmeter fail with unusual jmx property type
  - add variable sending example into Gatling documentation
  - fix results time and add widget for ApacheBenchmark
  - add `script: <jar-file>` capability to Gatling executor
+ - handle JMeter functions in `default-address` option
 
-## 1.6.7  <sup>23 aug 2016</sup>
+## 1.6.7 <sup>23 aug 2016</sup>
  - add `cloud` and `local` aliases 
  - add `detach` mode to Cloud Provisioning
  - add worker id to cloud log file names